(op-app): reset styles, separate Demo component
